fix speakers broekn tests
This PR fixes the broekn tests for speakers that was as a result of
unstable authentication system.

**Observation**

I also noticed that the backend is not formatted well and it's because
we don't run pre-commit before making a push.
A lot of file changes have been made on this PR due to pre-commit
formatting.